requirements.txt
# Core Framework
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
click==8.0.1
itsdangerous==2.0.1

# Database ORM
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
alembic==1.7.1

# Authentication
Flask-Login==0.5.0
oauthlib==3.1.1
requests-oauthlib==1.3.0

# Web Scraping and HTTP Requests
requests==2.26.0
beautifulsoup4==4.10.0
lxml==4.6.3

# PDF Generation
reportlab==3.6.1

# Image Processing
Pillow==8.3.2

# Development Tools
python-dotenv==0.19.0
pytest==6.2.5
pytest-flask==1.2.0

# Production Server
gunicorn==20.1.0

# Security
cryptography==3.4.8

# Scheduling and Automation
schedule==1.1.0
python-crontab==2.6.0

# JSON Processing
simplejson==3.17.6
